wiki-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
sqlalchemy-utils==0.41.2
asyncpg==0.29.0
psycopg2==2.9.10
alembic==1.12.1
pydantic[email,timezone]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4
minio==7.2.15
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0
mako==1.3.10
fastapi-cache2==0.2.2